+++ app/models/synchronization/adapter.rb
@@ -233,10 +233,7 @@ module CartoDB
         # if the geometry is MULTIPOINT we convert it to POINT
         if type == 'multipoint'
           user.db_service.in_database_direct_connection(statement_timeout: STATEMENT_TIMEOUT) do |user_database|
-            user_database.run("SELECT public.AddGeometryColumn('#{schema_name}', '#{table_name}','the_geom_simple',4326, 'GEOMETRY', 2);")
-            user_database.run(%Q{UPDATE #{qualified_table_name} SET the_geom_simple = ST_GeometryN(the_geom,1);})
-            user_database.run("SELECT DropGeometryColumn('#{schema_name}', '#{table_name}','the_geom');")
-            user_database.run(%Q{ALTER TABLE #{qualified_table_name} RENAME COLUMN the_geom_simple TO the_geom;})
+            user_database.run("UPDATE #{qualified_table_name} SET the_geom = ST_GeometryN(the_geom,1);")
           end
           type = 'point'
         end
@@ -244,10 +241,7 @@ module CartoDB
         # if the geometry is LINESTRING or POLYGON we convert it to MULTILINESTRING or MULTIPOLYGON
         if %w(linestring polygon).include?(type)
           user.db_service.in_database_direct_connection(statement_timeout: STATEMENT_TIMEOUT) do |user_database|
-            user_database.run("SELECT public.AddGeometryColumn('#{schema_name}', '#{table_name}','the_geom_simple',4326, 'GEOMETRY', 2);")
-            user_database.run(%Q{UPDATE #{qualified_table_name} SET the_geom_simple = ST_Multi(the_geom);})
-            user_database.run("SELECT DropGeometryColumn('#{schema_name}', '#{table_name}','the_geom');")
-            user_database.run(%Q{ALTER TABLE #{qualified_table_name} RENAME COLUMN the_geom_simple TO the_geom;})
+            user_database.run("UPDATE #{qualified_table_name} SET the_geom = ST_Multi(the_geom);")
 
             type = user_database[%Q{
               SELECT GeometryType(#{THE_GEOM})