+++ app/models/synchronization/member.rb
@@ -294,7 +294,8 @@ module CartoDB
         # last-modified time of the external data at the last synchronization) through the
         # downloader provided to the runner.
         # For connectors we need to pass it directly to the connector runner.
-        CartoDB::Importer2::ConnectorRunner.check_availability!(user)
+        provider_name = get_provider_name_from_id(service_item_id)
+        CartoDB::Importer2::ConnectorRunner.check_availability!(user, provider_name)
         CartoDB::Importer2::ConnectorRunner.new(
           service_item_id,
           user: user,
@@ -574,12 +575,23 @@ module CartoDB
       end
 
       def from_external_source?
-        ::ExternalDataImport.where(synchronization_id: self.id).first != nil
+        Carto::ExternalDataImport.where(synchronization_id: self.id).first != nil
       end
 
       attr_reader :repository
 
       attr_accessor :log_trace, :service_name, :service_item_id
+
+      private
+
+      def get_provider_name_from_id(service_item_id)
+        begin
+          connector_params = JSON.parse(service_item_id)
+          return connector_params['provider']
+        rescue StandardError
+          return nil
+        end
+      end
 
     end
   end