+++ spec/services/carto/do_licensing_service_spec.rb
@@ -11,7 +11,9 @@ describe Carto::DoLicensingService do
       dataset_id: @dataset_id,
       available_in: ['bq', 'bigtable'],
       price: 100,
-      expires_at: Time.new(2020, 9, 27, 8, 0, 0)
+      created_at: Time.new(2020, 9, 27, 7, 59, 0),
+      expires_at: Time.new(2021, 9, 27, 8, 0, 0),
+      status: 'active'
     }
   end
 
@@ -27,6 +29,8 @@ describe Carto::DoLicensingService do
     before(:each) do
       @central_mock = mock
       Cartodb::Central.stubs(:new).returns(@central_mock)
+      @service.stubs(:get_initial_sync_status).returns('unsynced')
+      @service.stubs(:get_entity_info).returns({})
     end
 
     after(:each) do
@@ -43,12 +47,15 @@ describe Carto::DoLicensingService do
       @central_mock.stubs(:create_do_datasets)
 
       bq_redis = [
-        { dataset_id: 'carto.abc.dataset1', expires_at: '2020-09-27 08:00:00 +0000' }
+        {
+          dataset_id: 'carto.abc.dataset1', created_at: '2020-09-27 07:59:00 +0000', expires_at: '2021-09-27 08:00:00 +0000',
+          status: 'active', available_in: ['bq', 'bigtable'], type: nil, estimated_size: 0, estimated_row_count: 0,
+          estimated_columns_count: 0, num_bytes: 0, sync_status: 'unsynced', unsyncable_reason: nil,
+          unsynced_errors: nil, sync_table: nil, sync_table_id: nil, synchronization_id: nil
+        }
       ].to_json
-      
-      bigtable_redis = [
-        { dataset_id: 'carto.abc.dataset1', expires_at: '2020-09-27 08:00:00 +0000' }
-      ].to_json
+
+      bigtable_redis = bq_redis
 
       @service.subscribe(@dataset)
       $users_metadata.hget(@redis_key, 'bq').should eq bq_redis
@@ -58,8 +65,12 @@ describe Carto::DoLicensingService do
     it 'allows to add more data in the same Redis key' do
       @central_mock.stubs(:create_do_datasets)
 
-      new_dataset = { dataset_id: 'carto.abc.dataset3', available_in: ['bq'], price: 300,
-          expires_at: '2020-09-27 08:00:00 +0000' }
+      new_dataset = {
+        dataset_id: 'carto.abc.dataset3',
+        available_in: ['bq'], price: 300,
+        expires_at: '2020-09-27 08:00:00 +0000',
+        status: 'active'
+      }
 
       @service.subscribe(@dataset)
       @service.subscribe(new_dataset)
@@ -75,6 +86,7 @@ describe Carto::DoLicensingService do
     before(:each) do
       @central_mock = mock
       Cartodb::Central.stubs(:new).returns(@central_mock)
+      @service.stubs(:get_initial_sync_status).returns('unsynced')
     end
 
     after(:each) do
@@ -88,6 +100,7 @@ describe Carto::DoLicensingService do
     end
 
     it 'removes the metadata from Redis' do
+      @service.stubs(:get_entity_info).returns(@dataset)
       @central_mock.stubs(:create_do_datasets)
       @central_mock.stubs(:remove_do_dataset)